Move components to common file
