+++ src/components/common/FileInput.css
@@ -45,7 +45,7 @@
   position: absolute;
   top: 14px;
   right: 14px;
-  background-image: url('../assets/close.svg');
+  background-image: url('../../assets/close.svg');
   background-position: center;
   background-repeat: no-repeat;
   padding: 0;
